Staging

+++ app/server.py
@@ -2,11 +2,13 @@ import traceback
 from flask import Flask, request, jsonify
 import minio
 from .engine import Template
+from .utils import download_minio_stream
 from typing import List, Dict
 import os
 import logging
 
 minio_client: minio.Minio = None
+
 db: Dict[str, Template] = {}
 template_folder = 'templates'
 
@@ -17,6 +19,10 @@ if not os.path.exists(template_folder):
 app = Flask(__name__)
 
 
+def make_name(filename: str) -> str:
+    return os.path.join(template_folder, filename)
+
+
 @app.route('/configure', methods=['POST'])
 def configure():
     global minio_client
@@ -25,12 +31,14 @@ def configure():
     host = data['host']
     access_key = data['access_key']
     pass_key = data['pass_key']
-    minio_client = minio.Minio(host, access_key, pass_key)
+    secure = data['secure']
+    try:
+        minio_client = minio.Minio(host, access_key, pass_key, secure=secure)
+        # checking that the instance is correct
+        minio_client.list_buckets()
+    except:
+        minio_client = None
     return jsonify({'error': False}), 200
-
-
-def make_name(filename: str) -> str:
-    return os.path.join(template_folder, filename)
 
 
 @app.route('/load_templates', methods=['POST'])
@@ -41,13 +49,11 @@ def load_template():
     for item in data:
         try:
             remote_bucket = item['bucket_name']
-            filename = item['template_name']
-            doc = minio_client.get_object(remote_bucket, filename)
-            name = make_name(filename)
-            with open(name, 'wb') as file_data:
-                for d in doc.stream(32*1024):
-                    file_data.write(d)
-            db[filename] = Template(name)
+            template_name = item['template_name']
+            doc = minio_client.get_object(remote_bucket, template_name)
+            name = make_name(template_name)
+            download_minio_stream(doc, name)
+            db[template_name] = Template(name)
             success.append(item)
         except:
             logging.error(traceback.format_exc())
@@ -64,7 +70,8 @@ def get_placeholders():
 
 @app.route('/publipost', methods=['POST'])
 def publipost():
-    body:Dict[str, object] = request.get_json()
+    body: Dict[str, object] = request.get_json()
+    output = None
     try:
 
         data: str = body['data']
@@ -76,16 +83,22 @@ def publipost():
         push_result = body.get('push_result', True)
         output = db[name].render(data)
 
-        if push_result :
+        if push_result:
             minio_client.fput_object(output_bucket, output_name, output)
             return jsonify({'error': False})
         else:
             # not used for now
             # should push the file back
-            return jsonify({'result':'OK'})
+            return jsonify({'result': 'OK'})
     except Exception as e:
         traceback.print_exc()
         return jsonify({'error': True}), 500
 
     finally:
-        os.remove(output)
+        if output is not None:
+            os.remove(output)
+
+
+@app.route('/live', methods=['GET'])
+def is_live():
+    return ('OK', 200) if minio_client is not None else ('KO', 402)
